pip-requirements.txt
django==4.1.3
python-telegram-bot==13.14